Remove inappropriate generic aliases for d050000l
The "D050000L" font was marked as an alias for "fantasy" but shouldn't be. According to the W3C CSS1 generic font-family definitions: "Fantasy fonts are primarily decorative or expressive fonts that contain decorative or expressive representations of characters. These do not include Pi or Picture fonts which do not represent actual characters."